Refactor config structures and loading logic
Moved configuration-related structs to a new config.h header for better modularity and maintainability. Updated config.c to use the new header and refactored the config loading logic into a reusable load_config function. Adjusted makefile to compile all .c files, and updated .gitignore for build outputs.
