+++ apps/sql-receptionist/.gitignore
@@ -1,5 +1,5 @@
+# Build output
 a.out
+*.exe
 
 test.py
-
-*.db
