+++ apps/sql-receptionist/makefile
@@ -2,8 +2,7 @@ CC = gcc
 CFLAGS = -I/mnt/c/Program\ Files/PostgreSQL/17/include
 LDFLAGS = -L/mnt/c/Program\ Files/PostgreSQL/17/lib -lpq -lcyaml
 TARGET = app.exe
-SRC = app.c
 
 app: $(SRC)
-	$(CC) $(CFLAGS) $(SRC) -o $(TARGET) $(LDFLAGS)
+	$(CC) $(CFLAGS) *.c -o $(TARGET) $(LDFLAGS)
 	./$(TARGET)
